+++ core/src/de/pcfreak9000/spaceawaits/registry/IBurnHandler.java
@@ -3,5 +3,5 @@ package de.pcfreak9000.spaceawaits.registry;
 import de.pcfreak9000.spaceawaits.item.Item;
 
 public interface IBurnHandler {
-    float getBurnTime(Item item);
+    int getBurnTime(Item item);
 }
